--- a/pixel-art-scaling/pixel_aa_no_gamma_correction.glslp
+++ b/pixel-art-scaling/pixel_aa_no_gamma_correction.glslp
@@ -1,0 +1,5 @@
+shaders = 1
+
+shader0 = shaders/pixel_aa/pixel_aa_no_gamma_correction.glsl
+filter_linear0 = true
+scale_type0 = viewport