--- a/pixel-art-scaling/shaders/pixel_aa/pixel_aa.glsl
+++ b/pixel-art-scaling/shaders/pixel_aa/pixel_aa.glsl
@@ -1,7 +1,7 @@
 #version 130
 
 /*
-    Pixel AA v1.6 by fishku
+    Pixel AA by fishku
     Copyright (C) 2023-2024
     Public domain license (CC0)
 
@@ -24,6 +24,7 @@
     subpixel anti-aliasing, results are identical to the "pixellate" shader.
 
     Changelog:
+    v1.7: Clean up, minor optimizations
     v1.6: Add "fast" version for low-end devices.
     v1.5: Optimize for embedded devices.
     v1.4: Enable subpixel sampling for all four pixel layout orientations,
@@ -36,7 +37,7 @@
 */
 
 // clang-format off
-#pragma parameter PIX_AA_SETTINGS "=== Pixel AA v1.6 settings ===" 0.0 0.0 1.0 1.0
+#pragma parameter PIX_AA_SETTINGS "=== Pixel AA v1.7 settings ===" 0.0 0.0 1.0 1.0
 #pragma parameter PIX_AA_SHARP "Pixel AA sharpening amount" 1.5 0.0 2.0 0.05
 #pragma parameter PIX_AA_GAMMA "Enable gamma-correct blending" 1.0 0.0 1.0 1.0
 #pragma parameter PIX_AA_SUBPX "Enable subpixel AA" 0.0 0.0 1.0 1.0
@@ -45,39 +46,23 @@
 
 #if defined(VERTEX)
 
-#if __VERSION__ >= 130
-#define COMPAT_VARYING out
-#define COMPAT_ATTRIBUTE in
-#define COMPAT_TEXTURE texture
-#else
-#define COMPAT_VARYING varying
-#define COMPAT_ATTRIBUTE attribute
-#define COMPAT_TEXTURE texture2D
-#endif
-
 #ifdef GL_ES
 #define COMPAT_PRECISION mediump
 #else
 #define COMPAT_PRECISION
 #endif
 
-COMPAT_ATTRIBUTE vec4 VertexCoord;
-COMPAT_ATTRIBUTE vec4 TexCoord;
-
-COMPAT_VARYING vec2 tx_coord;
-COMPAT_VARYING vec2 tx_per_px;
-COMPAT_VARYING vec2 tx_to_uv;
-
 uniform mat4 MVPMatrix;
-uniform COMPAT_PRECISION int FrameDirection;
-uniform COMPAT_PRECISION int FrameCount;
 uniform COMPAT_PRECISION vec2 OutputSize;
 uniform COMPAT_PRECISION vec2 TextureSize;
 uniform COMPAT_PRECISION vec2 InputSize;
 
-// compatibility #defines
-#define SourceSize vec4(TextureSize, 1.0 / TextureSize)
-#define OutSize vec4(OutputSize, 1.0 / OutputSize)
+in vec4 VertexCoord;
+in vec4 TexCoord;
+
+out vec2 tx_coord;
+out vec2 tx_per_px;
+out vec2 tx_to_uv;
 
 void main() {
   gl_Position = MVPMatrix * VertexCoord;
@@ -99,34 +84,14 @@ precision mediump float;
 #define COMPAT_PRECISION
 #endif
 
-#if __VERSION__ >= 130
-#define COMPAT_VARYING in
-#define COMPAT_TEXTURE texture
-out COMPAT_PRECISION vec4 FragColor;
-#else
-#define COMPAT_VARYING varying
-#define FragColor gl_FragColor
-#define COMPAT_TEXTURE texture2D
-#endif
-
-uniform COMPAT_PRECISION int FrameDirection;
-uniform COMPAT_PRECISION int FrameCount;
-uniform COMPAT_PRECISION vec2 OutputSize;
-uniform COMPAT_PRECISION vec2 TextureSize;
-uniform COMPAT_PRECISION vec2 InputSize;
 uniform COMPAT_PRECISION int Rotation;
 uniform sampler2D Texture;
 
-COMPAT_VARYING vec2 tx_coord;
-COMPAT_VARYING vec2 tx_per_px;
-COMPAT_VARYING vec2 tx_to_uv;
+in vec2 tx_coord;
+in vec2 tx_per_px;
+in vec2 tx_to_uv;
 
-// compatibility #defines
-#define Source Texture
-#define vTexCoord TEX0.xy
-
-#define SourceSize vec4(TextureSize, 1.0 / TextureSize)
-#define OutSize vec4(OutputSize, 1.0 / OutputSize)
+out COMPAT_PRECISION vec4 FragColor;
 
 // Similar to smoothstep, but has a configurable slope at x = 0.5.
 // Original smoothstep has a slope of 1.5 at x = 0.5
@@ -137,10 +102,8 @@ vec2 slopestep(vec2 edge0, vec2 edge1, vec2 x, float slope) {
   return o - 0.5 * s * pow(2.0 * (o - s * x), vec2(slope));
 }
 
-float to_lin(float x) { return pow(x, 2.2); }
 vec3 to_lin(vec3 x) { return pow(x, vec3(2.2)); }
-
-float to_srgb(float x) { return pow(x, 1.0 / 2.2); }
+vec4 to_lin(vec4 x) { return pow(x, vec4(2.2)); }
 vec3 to_srgb(vec3 x) { return pow(x, vec3(1.0 / 2.2)); }
 
 // Function to get a pixel value, taking into consideration possible subpixel
@@ -168,64 +131,58 @@ vec4 pixel_aa(sampler2D tex, vec2 tx_per_px, vec2 tx_to_uv, vec2 tx_coord,
     vec2 period, phase, offset;
 
     if (gamma_correct) {
+      vec4 samples;
       // Red
       period = floor(tx_coord - sub_tx_offset - 0.5);
       phase = tx_coord - sub_tx_offset - 0.5 - period;
       offset = slopestep(sharp_lb, sharp_ub, phase, sharpness_lower);
-      res.r = to_srgb(mix(
-          mix(to_lin(COMPAT_TEXTURE(tex, (period + 0.5) * tx_to_uv).r),
-              to_lin(
-                  COMPAT_TEXTURE(tex, (period + vec2(1.5, 0.5)) * tx_to_uv).r),
-              offset.x),
-          mix(to_lin(
-                  COMPAT_TEXTURE(tex, (period + vec2(0.5, 1.5)) * tx_to_uv).r),
-              to_lin(COMPAT_TEXTURE(tex, (period + 1.5) * tx_to_uv).r),
-              offset.x),
-          offset.y));
+      samples = vec4(texture(tex, (period + 0.5) * tx_to_uv).r,
+                     texture(tex, (period + vec2(1.5, 0.5)) * tx_to_uv).r,
+                     texture(tex, (period + vec2(0.5, 1.5)) * tx_to_uv).r,
+                     texture(tex, (period + 1.5) * tx_to_uv).r);
+      samples = to_lin(samples);
+      res.r = mix(mix(samples.x, samples.y, offset.x),
+                  mix(samples.z, samples.w, offset.x), offset.y);
       // Green
       period = floor(tx_coord - 0.5);
       phase = tx_coord - 0.5 - period;
       offset = slopestep(sharp_lb, sharp_ub, phase, sharpness_lower);
-      res.g = to_srgb(mix(
-          mix(to_lin(COMPAT_TEXTURE(tex, (period + 0.5) * tx_to_uv).g),
-              to_lin(
-                  COMPAT_TEXTURE(tex, (period + vec2(1.5, 0.5)) * tx_to_uv).g),
-              offset.x),
-          mix(to_lin(
-                  COMPAT_TEXTURE(tex, (period + vec2(0.5, 1.5)) * tx_to_uv).g),
-              to_lin(COMPAT_TEXTURE(tex, (period + 1.5) * tx_to_uv).g),
-              offset.x),
-          offset.y));
+      samples = vec4(texture(tex, (period + 0.5) * tx_to_uv).g,
+                     texture(tex, (period + vec2(1.5, 0.5)) * tx_to_uv).g,
+                     texture(tex, (period + vec2(0.5, 1.5)) * tx_to_uv).g,
+                     texture(tex, (period + 1.5) * tx_to_uv).g);
+      samples = to_lin(samples);
+      res.g = mix(mix(samples.x, samples.y, offset.x),
+                  mix(samples.z, samples.w, offset.x), offset.y);
       // Blue
       period = floor(tx_coord + sub_tx_offset - 0.5);
       phase = tx_coord + sub_tx_offset - 0.5 - period;
       offset = slopestep(sharp_lb, sharp_ub, phase, sharpness_lower);
-      res.b = to_srgb(mix(
-          mix(to_lin(COMPAT_TEXTURE(tex, (period + 0.5) * tx_to_uv).b),
-              to_lin(
-                  COMPAT_TEXTURE(tex, (period + vec2(1.5, 0.5)) * tx_to_uv).b),
-              offset.x),
-          mix(to_lin(
-                  COMPAT_TEXTURE(tex, (period + vec2(0.5, 1.5)) * tx_to_uv).b),
-              to_lin(COMPAT_TEXTURE(tex, (period + 1.5) * tx_to_uv).b),
-              offset.x),
-          offset.y));
+      samples = vec4(texture(tex, (period + 0.5) * tx_to_uv).b,
+                     texture(tex, (period + vec2(1.5, 0.5)) * tx_to_uv).b,
+                     texture(tex, (period + vec2(0.5, 1.5)) * tx_to_uv).b,
+                     texture(tex, (period + 1.5) * tx_to_uv).b);
+      samples = to_lin(samples);
+      res.b = mix(mix(samples.x, samples.y, offset.x),
+                  mix(samples.z, samples.w, offset.x), offset.y);
+
+      res = to_srgb(res);
     } else {
       // Red
       period = floor(tx_coord - sub_tx_offset - 0.5);
       phase = tx_coord - sub_tx_offset - 0.5 - period;
       offset = slopestep(sharp_lb, sharp_ub, phase, sharpness_lower);
-      res.r = COMPAT_TEXTURE(tex, (period + 0.5 + offset) * tx_to_uv).r;
+      res.r = texture(tex, (period + 0.5 + offset) * tx_to_uv).r;
       // Green
       period = floor(tx_coord - 0.5);
       phase = tx_coord - 0.5 - period;
       offset = slopestep(sharp_lb, sharp_ub, phase, sharpness_lower);
-      res.g = COMPAT_TEXTURE(tex, (period + 0.5 + offset) * tx_to_uv).g;
+      res.g = texture(tex, (period + 0.5 + offset) * tx_to_uv).g;
       // Blue
       period = floor(tx_coord + sub_tx_offset - 0.5);
       phase = tx_coord + sub_tx_offset - 0.5 - period;
       offset = slopestep(sharp_lb, sharp_ub, phase, sharpness_lower);
-      res.b = COMPAT_TEXTURE(tex, (period + 0.5 + offset) * tx_to_uv).b;
+      res.b = texture(tex, (period + 0.5 + offset) * tx_to_uv).b;
     }
 
     return vec4(res, 1.0);
@@ -250,38 +207,29 @@ vec4 pixel_aa(sampler2D tex, vec2 tx_per_px, vec2 tx_to_uv, vec2 tx_coord,
     if (gamma_correct) {
       return vec4(
           to_srgb(mix(
-              mix(to_lin(COMPAT_TEXTURE(tex, (period + 0.5) * tx_to_uv).rgb),
+              mix(to_lin(texture(tex, (period + 0.5) * tx_to_uv).rgb),
                   to_lin(
-                      COMPAT_TEXTURE(tex, (period + vec2(1.5, 0.5)) * tx_to_uv)
-                          .rgb),
+                      texture(tex, (period + vec2(1.5, 0.5)) * tx_to_uv).rgb),
                   offset.x),
               mix(to_lin(
-                      COMPAT_TEXTURE(tex, (period + vec2(0.5, 1.5)) * tx_to_uv)
-                          .rgb),
-                  to_lin(COMPAT_TEXTURE(tex, (period + 1.5) * tx_to_uv).rgb),
+                      texture(tex, (period + vec2(0.5, 1.5)) * tx_to_uv).rgb),
+                  to_lin(texture(tex, (period + 1.5) * tx_to_uv).rgb),
                   offset.x),
               offset.y)),
           1.0);
     } else {
-      return COMPAT_TEXTURE(tex, (period + 0.5 + offset) * tx_to_uv);
+      return texture(tex, (period + 0.5 + offset) * tx_to_uv);
     }
   }
 }
 
-#ifdef PARAMETER_UNIFORM
 uniform COMPAT_PRECISION float PIX_AA_SHARP;
 uniform COMPAT_PRECISION float PIX_AA_GAMMA;
 uniform COMPAT_PRECISION float PIX_AA_SUBPX;
 uniform COMPAT_PRECISION float PIX_AA_SUBPX_ORIENTATION;
-#else
-#define PIX_AA_SHARP 1.5
-#define PIX_AA_GAMMA 1.0
-#define PIX_AA_SUBPX 0.0
-#define PIX_AA_SUBPX_ORIENTATION 0.0
-#endif
 
 void main() {
-  FragColor = pixel_aa(Source, tx_per_px, tx_to_uv, tx_coord, PIX_AA_SHARP,
+  FragColor = pixel_aa(Texture, tx_per_px, tx_to_uv, tx_coord, PIX_AA_SHARP,
                        PIX_AA_GAMMA > 0.5, PIX_AA_SUBPX > 0.5,
                        int(PIX_AA_SUBPX_ORIENTATION), Rotation);
 }