--- a/pixel-art-scaling/shaders/pixel_aa/pixel_aa_fast.glsl
+++ b/pixel-art-scaling/shaders/pixel_aa/pixel_aa_fast.glsl
@@ -27,10 +27,10 @@ out PREC_MED vec2 px_per_tx;
 out PREC_MED vec2 tx_to_uv;
 
 void main() {
-  gl_Position = MVPMatrix * VertexCoord;
-  tx_coord = TexCoord.xy * TextureSize - 0.5;
-  px_per_tx = OutputSize / InputSize;
-  tx_to_uv = 1.0 / TextureSize;
+    gl_Position = MVPMatrix * VertexCoord;
+    tx_coord = TexCoord.xy * TextureSize - 0.5;
+    px_per_tx = OutputSize / InputSize;
+    tx_to_uv = 1.0 / TextureSize;
 }
 
 #elif defined(FRAGMENT)
@@ -48,14 +48,14 @@ in PREC_MED vec2 tx_to_uv;
 out PREC_LOW vec4 FragColor;
 
 void main() {
-  PREC_MED vec2 period;
-  PREC_MED vec2 phase = modf(tx_coord, period);
+    PREC_MED vec2 period;
+    PREC_MED vec2 phase = modf(tx_coord, period);
 
-  PREC_MED vec2 t = clamp((phase - 0.5) * px_per_tx + 0.5, 0.0, 1.0);
-  PREC_MED vec2 offset = t * t * (3.0 - 2.0 * t);
+    PREC_MED vec2 t = clamp((phase - 0.5) * px_per_tx + 0.5, 0.0, 1.0);
+    PREC_MED vec2 offset = t * t * (3.0 - 2.0 * t);
 
-  PREC_LOW vec3 res = texture(Texture, (period + 0.5 + offset) * tx_to_uv).rgb;
-  FragColor = vec4(sqrt(res), 1.0);
+    PREC_LOW vec3 res = texture(Texture, (period + 0.5 + offset) * tx_to_uv).rgb;
+    FragColor = vec4(sqrt(res), 1.0);
 }
 
 #endif