Clean up, heavily optimize.

On low-end RK3326 device, improvements are as follows:
pixel_aa.glslp:
Improve from 23 to 10 ms with default params (~80% improvement)
31 -> 21 ms with sub-pixel sampling (~45% improvement)

pixel_aa_fast.glslp:
9.2 -> 8.5 ms (~30% improvement)